Add morse_button service and update morse_chat with IPC
Added the morse_button service which interacts with the
button driver, letting the user record a morse message
by using the 2 buttons on the back of the micro:bit.

The service needs to first enter sending mode, by holding
first the Left Button, then the Right button, then letting
go of the Left and then the Right button. Altering this
sequence will not enter sending/recording mode.

After entering sending mode, the user can use the Left button
to record a '.' character and the Right button for a '-'.

The sending mode is exited when the buffer is full and the
characters are automatically sent towards the POST endpoint,
or an early exit can be achieved by doing the same sequence
required to enter the sending mode.

Signed-off-by: Andrei Cherechesu <andrei.cherechesu@gmail.com>