Add mutual exclusion mechanism for network
Added a global variable read by the morse_chat app
from the buffer that the morse_button app is using for
passing data to it.

If the value read is 0xff, then
the morse_button app is in 'recording' or 'sending' state,
and the network is therefore locked until a POST request
is completed.

Whenever the callback is entered after a completed transmission,
the global flag is set to false and the network is unlocked.

Signed-off-by: Andrei Cherechesu <andrei.cherechesu@gmail.com>